vinyl: set run_size_ratio to 3.5 by default
Closes #2038

+++ src/box/key_def.h
@@ -200,7 +200,7 @@ struct key_opts {
 	 * the LSM tree is run_size_ratio times larger than
 	 * previous one.
 	 */
-	int64_t run_size_ratio;
+	double run_size_ratio;
 	/**
 	 * LSN from the time of index creation.
 	 */
